# src/policy-price-manager/public/swagger.json
{
	"components": {
...
		"securitySchemes": {}
	},
	"info": {
		"title": "policy-price-manager",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/establishmment/{establishmentId}": {
			"get": {
				"operationId": "FindByEstablishment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"httpCode": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"httpCode"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "establishmentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}